+++ Logos.h
@@ -1,0 +1,10 @@
+#ifndef LOGOS_H
+#define LOGOS_H
+
+#include <Arduino.h>
+
+extern const unsigned char checkMark[];
+extern const unsigned char warningMark[];
+
+
+#endif